Restructure app to match Angular tutorial
I was continuing to have build problems, which are apparently due to having
an unexpected application structure. I've significantly refactored in this commit
to match the Angular tutorial
(https://github.com/angular/angular.dart.tutorial/tree/master/Chapter_06). All of
a sudden, the build works and I can run this in a regular browser!